+++ src/core/component.js
