+++ src/core/register-component.js
